Add files via upload
In a creative exploration of web development, I designed and developed a three-page e-commerce website themed around the concept of 'Internet in the 1980s', focusing on radios. This project, built with HTML, CSS, and JavaScript, showcased a vintage aesthetic with modern web functionalities. It featured a home page, product catalog, and a contact form, offering a unique blend of retro design elements with contemporary e-commerce features. This project not only honed my front-end development skills but also challenged my creativity in integrating thematic design with user-friendly website architecture.